[ADD][MAGN-0042] woo_commerce_magnety: fix min_quantity error

+++ woo_commerce_magnety/models/product_ept.py
@@ -226,7 +226,7 @@ class WooProductTemplateEpt(models.Model):
                         min_quantity = get_value(meta_data, f'_bulkdiscount_quantity_{i}')
                         discount = get_value(meta_data, f'_bulkdiscount_discount_fixed_{i}')
                         if min_quantity and discount:
-                            min_quantity = int(min_quantity)
+                            min_quantity = float(min_quantity)
                             discount = float(discount)
                             bulk_price = base_price.fixed_price - discount
                             exist_price = Price.search([
